first commit
KDDTest.csv is the test data(Set1 2 3 are temporary files). 20 Percent Training Set.csv is the training data(for K-NN).
add the dataset and my scripts(python files). run preprocess.py to process the data(training data or test data, choosed in parameters)
test.py is a script in development. I try to calculate the nearest distance of all 20000+ rows in test dataset using test function, which is really slow.